requirements.txt
aniso8601==7.0.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
deprecation==2.1.0
ecdsa==0.18.0
Flask==2.3.2
Flask-GraphQL==2.0.1
Flask-SQLAlchemy==3.0.5
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
packaging==23.1
promise==2.3
protobuf==4.21.12
pyasn1==0.5.0
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
Rx==1.6.3
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.49
urllib3==2.0.4
Werkzeug==2.3.6
